added args+

+++ src/common/traits.rs
@@ -7,6 +7,6 @@ pub trait ConnectionFn {
     where
         Self: 'a;
 
-    fn execute(&self, q: &str, fetch_num: usize) -> Result<(Vec<ColDesc>, Self::Cursor<'_>), Error>;
-    fn fetch(&self, c: &mut Self::Cursor<'_>, fetch_num: usize) -> Result<Vec<Vec<String>>, Error>;
+    fn execute(&self, q: &str, fetch_num: i32) -> Result<(Vec<ColDesc>, Self::Cursor<'_>), Error>;
+    fn fetch(&self, c: &mut Self::Cursor<'_>, fetch_num: i32) -> Result<Vec<Vec<String>>, Error>;
 }
